Implement caching for rule results
Theoretically, it brings parsing time to liner time relative to the
input. Cache is implemented per-parsing state, rule and position. If we
already computed the rule for the same parsing at the same position, we
return previously computed result and restore tokenizer pointer to the
position of the last match and skip the matching.